requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
Cython==0.29.21
Django==3.1.8
django-crispy-forms==1.10.0
django-pandas==0.6.2
django-widget-tweaks==1.4.12
dlib==19.21.1
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==2.10
imutils==0.5.3
joblib==0.17.0
kiwisolver==1.3.1
matplotlib==3.3.3
mypy-extensions==0.4.3
mysqlclient==2.1.0
numpy==1.19.3
opencv-python==4.4.0.46
pandas==1.1.5
pathspec==0.8.1
Pillow==8.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
regex==2020.11.13
requests==2.25.1
scikit-learn==0.21.2
scipy==1.5.4
seaborn==0.11.0
six==1.15.0
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
